Parse: Revise algo. used to compute durations.
This patch improves the robustness of the TimeTracker class
that is used to match records from the same job. The original
code was very vulnerable to messages recorded out of order,
leading to inaccurate data.

This patch does slow down record parsing, but accurate results
are more important than fast results.